refactor: moves the job macro implementation out of lib.rs into job.rs, adds mod job to lib.rs and makes job_macro_impl pub(crate)
